pyproject.toml
[tool.poetry]
name = "OneLauncher"
version = "2.0.dev"
description = "The OneLauncher to rule them all"
readme = "README.md"
authors = ["June Stepp <contact@junestepp.me>"]
license = "GPL-3.0-or-later"
repository = "https://github.com/JuneStepp/OneLauncher"
packages = [{ include = "onelauncher", from = "src" }]
keywords = ["LOTRO", "DDO", "launcher", "addon-manager", "custom-launcher"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Topic :: Games/Entertainment :: Role-Playing",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
PySide6-Essentials = "^6.6.1"

# Keyring packages
keyring = "^24.2.0"
cryptography = "^42.0.5"
SecretStorage = { version = "^3.3.3", markers = "sys_platform == 'linux'" }
# Should only use with KWallet versions < 5.97. See https://bugs.kde.org/show_bug.cgi?id=458318.
dbus-python = { version = "^1.3.2", markers = "sys_platform == 'linux'", optional = true }

defusedxml = "^0.7.1"
vkbeautify = "^0.3.0"
rtoml = "^0.9.0"
platformdirs = "^4.2.0"
xmlschema = "^3.2.0"
feedparser = "^6.0.10"
Babel = "^2.12.1"
zeep = "^4.2.1"
cachetools = "^5.3.1"
trio = "^0.25.0"
httpx = "^0.26.0"
asyncache = { git = "https://github.com/JuneStepp/asyncache.git" }
attrs = "^23.2.0"
cattrs = { version = "^23.2.3", extras = ["tomlkit"] }
packaging = "^24.0"
typer = "^0.12.3"

[tool.poetry.group.dev.dependencies]
Nuitka = "^2.1.5"
pytest = "^7.4.2"
mypy = "^1.9.0"
types-cachetools = "^5.3.0"
trio-typing = { extras = ["mypy"], version = "^0.10.0" }
ruff = "^0.3.7"

[tool.poetry.scripts]
onelauncher = "onelauncher.new_cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311" # TODO: Remove once Poetry supports [project] section
extend-exclude = ["*_uic.py"] # Ignore autogenerated UI files

[tool.ruff.lint]
select = [
    "E4",    # pycodestyle
    "E7",
    "E9",
    "F",     # Pyflakes
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "I",     # isort,
    "RUF",   # Ruff
    "S",     # flake8-bandit
    "ASYNC", # flake8-async
    "TRIO",  #flake8-trio,
    "ANN",   #flake8-annotations
    "A",     #flake8-builtins
    "CPY",   # flake8-copyright
    "FA",    # flake8-future-annotations
    "T20",   # flake8-print
    "FIX",   # flake8-fixme
    "ERA",   # eradicate
    "PL",    # Pylint

]
ignore = [
    "ANN101",  # missing-type-self
    "PLR0913", # too-many-arguments
]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["S101"] # assert

[tool.ruff.lint.flake8-annotations]
mypy-init-return = true

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

[tool.mypy]
plugins = "trio_typing.plugin"

[[tool.mypy.overrides]]
module = [
    "feedparser", # TODO Future version will have typing, 
    # once https://github.com/kurtmckee/feedparser/blob/develop/changelog.d/20210801_133300_palfrey_typing.rst
    # makes it to a release
]
strict = true
